Sets appropriate title and exhortation text to popup
Change ScriptureHolder into SceneRelay, and add a boolean win field to it with a getter and a setter.

Used the SceneRelay class to communicate who won to the results window. Upon receiving this data, the result window sets the appropriate text.

It might be better to clean up the Hangman controller and separate things into more methods