New Scene Switching System
Instead of keeping the same stage and switching scenes, the setup, game, and result window are all separate stages now.

SceneRelay now has a Hashmap<String, Stage>. As soon as possible in each controller, the stage is logged in the hashmap for access. So, whenever we want to change the view, we simply retrieve the stage we want, call Stage.show(), then hide our own stage.

This way, we can keep state even though the stage is not visible.

Now that we've done this, we need to do a lot of cleaning up from the old way. But for now, everything is functional as far as replays.