Webscraper Grammar #3
Add a line for detached periods.

Update the calls to fix 'floating punctuation" instead of " . " -> ". "
we have " ." -> "."

This identifies the real problem, which is when punctuation is not abutted to the right side of a word. When that happens, simply eliminate the space in between them. Fixes cases with stacked punctuation. Old solution would ignore this error: "This the end of my sentence !"